Add a "validator box" job, separate validation and post-processing
Use the webhook version as a job to have a permanent box to attach to for some validation debugging.
Split the validation steps in two files to make it easier to deploy the xml-check + rules or the svrl + json post-processing independently.
Add an SVG with draw.io source to document the processing steps.